src/requirements.txt
boto3
opencv-python
Pillow
imutils
requests